NSF interface overhaul.
NSF player options added.
* Track select
* Pause
* Restart song
* Muting audio / expansion audio.
* Fixed some compiler warnings.
* Fixed some Camera mapper stuff
* Fixed FDS image detection. 0 byte file are not detected as FDS images causing a crash.
* Audio buffers reserving RAM. Improves memory allocation load.
* Virtual CPU crash also dumps the 2a03's stack.
* NSF player computes time not from SDL_GetTicks but based on how many virtual 2a03 ticks has been passed.
*

+++ 3rdparty/hqx/HQ2x.cpp
@@ -71,7 +71,7 @@ uint32_t *HQ2x::resize(
 
 		// adjusts the previous and next line pointers
 		if (row > 0)
-			previous = -width;
+			previous = -(int)width;
 		else
 		{
 			if (wrapY)
@@ -84,7 +84,7 @@ uint32_t *HQ2x::resize(
 		else
 		{
 			if (wrapY)
-				next = -(width * (height - 1));
+				next = -((int)width * ((int)height - 1));
 			else
 				next = 0;
 		}
